Make is_dse_job a property of TestDefinition
